+++ tools/server/boot.js
@@ -6,7 +6,7 @@ var _ = require('underscore');
 var sourcemap_support = require('source-map-support');
 
 // This code is duplicated in tools/main.js.
-var MIN_NODE_VERSION = 'v0.10.29';
+var MIN_NODE_VERSION = 'v0.10.33';
 
 if (require('semver').lt(process.version, MIN_NODE_VERSION)) {
   process.stderr.write(